enhance/migrated from plain react to nextjs
