+++ .husky/pre-push
@@ -1,0 +1,19 @@
+#!/usr/bin/env sh
+. "$(dirname -- "$0")/_/husky.sh"
+
+branch="$(git rev-parse --abbrev-ref HEAD)"
+
+if [ "$branch" = "main" ]; then
+  echo "Pushing to main branch is not allowed 😜"
+  exit 1
+fi
+
+if [ "$branch" = "stage" ]; then
+  echo "Pushing to stage branch is not allowed 😜"
+  exit 1
+fi
+
+echo 'Building the project ⏳'
+
+npx next build
+
