+++ .prettierrc.json
@@ -1,8 +1,8 @@
 {
-  "printWidth": 60,
-  "tabWidth": 2,
   "semi": true,
-  "singleQuote": false,
+  "tabWidth": 2,
+  "printWidth": 80,
+  "endOfLine": "auto",
   "trailingComma": "all",
-  "bracketSpacing": true
+  "plugins": ["prettier-plugin-tailwindcss"]
 }
