--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-  "editor.fontSize": 26,
-  "terminal.integrated.fontSize": 20,
-  "prettier.printWidth": 60
-}
