--- index.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html class="scroll-smooth" lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" href="/favicon.ico" />
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="/apple-touch-icon.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="32x32"
-      href="/favicon-32x32.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="16x16"
-      href="/favicon-16x16.png"
-    />
-    <link rel="manifest" href="/site.webmanifest" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0"
-    />
-    <title>Pezhman Ghavami - Front-End Developer</title>
-    <meta
-      name="twitter:card"
-      content="summary_large_image"
-    />
-    <meta
-      name="twitter:title"
-      content="Pezhman Ghavami - Front-End Developer"
-    />
-    <meta
-      name="twitter:description"
-      content="Pezhman Ghavami's Portfolio website"
-    />
-    <meta
-      name="twitter:image"
-      content="https://res.cloudinary.com/drsgyshsf/image/upload/v1680355891/my-portfolio/og-and-tc/pezhmanghavami-tc_spuarz.png"
-    />
-    <meta
-      property="og:title"
-      content="Pezhman Ghavami - Front-End Developer"
-    />
-    <meta
-      property="og:description"
-      content="Pezhman Ghavami's Portfolio website"
-    />
-    <meta
-      property="og:image"
-      content="https://res.cloudinary.com/drsgyshsf/image/upload/v1680355890/my-portfolio/og-and-tc/pezhmanghavami-og_vl5x4l.png"
-    />
-  </head>
-  <body class="bg-neutral-900 text-neutral-300">
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
