+++ package.json
@@ -1,30 +1,44 @@
 {
   "name": "my-portfolio",
+  "version": "0.1.0",
   "private": true,
-  "version": "1.0.0",
-  "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview",
-    "update": "ncu -u && npm i"
+    "postinstall": "husky",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "npm run lint:ts && npm run lint:next && npm run lint:eslint && npm run lint:stylelint && npm run format",
+    "lint:next": "next lint",
+    "lint:ts": "tsc -p tsconfig.json --noEmit",
+    "lint:eslint": "eslint .",
+    "format": "prettier --write .",
+    "update": "ncu -u && npm i",
+    "prepare": "husky"
   },
   "dependencies": {
-    "npm-check-updates": "^16.10.12",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "next": "14.2.5",
+    "react": "^18",
+    "react-dom": "^18"
   },
   "devDependencies": {
-    "@types/react": "^18.2.12",
-    "@types/react-dom": "^18.2.5",
-    "@vitejs/plugin-react": "^4.0.0",
-    "autoprefixer": "^10.4.14",
-    "postcss": "^8.4.24",
-    "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.3.0",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.1.3",
-    "vite": "^4.3.9",
-    "vite-plugin-svgr": "^3.2.0"
+    "@svgr/webpack": "^8.1.0",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "14.2.5",
+    "npm-check-updates": "^16.14.20",
+    "postcss": "^8",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5",
+    "@typescript-eslint/eslint-plugin": "^7.12.0",
+    "@typescript-eslint/parser": "^7.12.0",
+    "eslint-config-prettier": "^9.1.0",
+    "prettier": "^3.3.1",
+    "prettier-plugin-tailwindcss": "^0.6.2",
+    "tailwind-merge": "^2.3.0",
+    "clsx": "^2.1.1",
+    "husky": "^9.0.11",
+    "lint-staged": "^15.2.5"
   }
 }
