+++ svgr.d.ts
@@ -1,0 +1,10 @@
+declare module "*.svg" {
+  import { FC, SVGProps } from "react";
+  const content: FC<SVGProps<SVGElement>>;
+  export default content;
+}
+
+declare module "*.svg?url" {
+  const content: any;
+  export default content;
+}
