--- vite.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import svgr from "vite-plugin-svgr";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react(), svgr()],
-});
